+++ pricing/src/epilot/sdkconfiguration.py
@@ -22,9 +22,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.2'
-    gen_version: str = '2.194.1'
-    user_agent: str = 'speakeasy-sdk/python 2.1.2 2.194.1 1.0.0 epilot-pricing'
+    sdk_version: str = '2.1.3'
+    gen_version: str = '2.195.2'
+    user_agent: str = 'speakeasy-sdk/python 2.1.3 2.195.2 1.0.0 epilot-pricing'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: